backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Pydantic for data validation and settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Neo4j driver
neo4j>=5.15.0

# HTTP client for API calls
httpx>=0.26.0

# JWT token verification for Clerk
PyJWT>=2.8.0
cryptography>=41.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
hypothesis>=6.92.0

# Environment variables
python-dotenv>=1.0.0
